eigth commit

+++ helpers/productHelpers.js
@@ -66,6 +66,13 @@ const addressInfo = mongoose.model('address',addressSchema)
 //Schema for coupons
 const couponInfo = require("./couponSchema")
 
+//Schema for wishlist
+const wishlistSchema = new mongoose.Schema({
+    user:String,
+    products:Array
+})
+//collection for wishList
+const wishlistInfo = mongoose.model('wishlist',wishlistSchema)
  
 module.exports={
   addProducts  : (product,id1,id2,id3)=>{
@@ -360,12 +367,14 @@ module.exports={
                 })
         })
     },
-    deleteItem:(id)=>{
+    deleteCartItem:(id)=>{
+       
         return new Promise((resolve,rejection)=>{
             let userId = id.userId
             let proId = id.proId
             cartInfo.updateOne({user:userId},{$pull:{products:{items:objectId(proId)}}}).then((res)=>{
-               
+           
+               console.log(res)
                 resolve(true)
             })
         })
@@ -469,9 +478,9 @@ getProductList:(userId)=>{
      
     })
 },
-placeOrder:(address,products,total,paymentMethod,userId)=>{
+placeOrder:(address,products,total,paymentMethod,userId,flag)=>{
   
-  
+
   
     return new Promise((resolve,reject)=>{
         
@@ -496,10 +505,13 @@ placeOrder:(address,products,total,paymentMethod,userId)=>{
         
                 resolve(details)
             }
+
+    if(flag!=1){
          cartInfo.deleteOne({user:userId}).then((res)=>{
            
-             console.log(res,"++++++++++++++++++++++++++++")
+             
          })
+        }
         })
        
     })
@@ -548,16 +560,14 @@ cancelOrder:(orderId)=>{
     console.log(ordersId)
     return new Promise(async(resolve,reject)=>{
        await orderInfo.updateOne({_id:ordersId},{$set:{status:"cancelled"}}).then((response)=>{
-        console.log("updated status")
-        console.log(response)
+      resolve(response)
     })
     })
 },
 getAllOrders:()=>{
     return new Promise((resolve,reject)=>{
         orderInfo.find().then((res)=>{
-            console.log("get all orders")
-            console.log(res)
+         
             resolve(res)
         })
     })
@@ -567,11 +577,9 @@ mangeOrder:(Id,status)=>{
     
     return new Promise(async(resolve,reject)=>{
     let orderId = objectId(Id);
-     console.log("order id in mangemfjsd")
-    console.log(orderId)
-    console.log(status)
+
       await orderInfo.updateOne({_id:orderId},{$set:{status:status}}).then((response)=>{
-        console.log(response)
+       
         resolve(response)
       })
       
@@ -580,12 +588,10 @@ mangeOrder:(Id,status)=>{
    
 },
 addAddress:(body,userId,addressId)=>{
-    console.log("ivde ondo  ")
+    
  return new Promise(async(resolve,reject)=>{
     
-     console.log("body in address")
-     console.log(userId)
-     console.log(body)
+    
      let addressObj ={
          addressId:addressId,
          name:body.name,
@@ -620,19 +626,18 @@ getAddress:(userId,aid)=>{
     return new Promise(async(resolve,reject)=>{
         // addressID = objectId(addressId)
         let addressFeild = "address.addressId"
-        console.log("get adress aid")
+  
        
 
 
-        console.log(userId)
+        
         let addressId ="address.addressId"
         let singleAddress =await addressInfo.aggregate([
             {$match:{userId:userId}},
             {$unwind:"$address"},
             {$match:{[addressFeild]:aid.addressId}}
         ])
-        console.log("aggregated single address")
-        console.log(singleAddress)
+        
         resolve(singleAddress[0])
         
          
@@ -646,15 +651,14 @@ getSingleAddress:(userId,aid)=>{
        
 
 
-        console.log(userId)
+        
         let addressId ="address.addressId"
         let singleAddress =await addressInfo.aggregate([
             {$match:{userId:userId}},
             {$unwind:"$address"},
             {$match:{[addressFeild]:aid}}
         ])
-        console.log("aggregated single address")
-        console.log(singleAddress)
+      
         resolve(singleAddress[0])
         
          
@@ -669,7 +673,7 @@ ediAddress:(body,userId)=>{
     let pincode = body.pincode
     let addressname = body.addressname
     let phone = body.phone
-    console.log(name,email,address)
+   
     return new Promise((resolve,reject)=>{
         let updatedName= "address.$.name"
         let updatedEmail =  "address.$.email"
@@ -691,7 +695,7 @@ ediAddress:(body,userId)=>{
             }
              
             ).then((res)=>{
-                console.log(res)
+               
                 resolve(res)
             })
           
@@ -726,11 +730,10 @@ ediAddress:(body,userId)=>{
 deleteAdress:(body)=>{
     let userId = body.userID
     let Aid = body.aid
-    console.log(userId)
-    console.log(Aid)
+
     return new Promise((resolve,reject)=>{
         addressInfo.updateOne({userId:userId},{$pull:{address:{addressId:Aid}}}).then((res)=>{
-            console.log("response after delete")
+           
             resolve(res)
         })
     })
@@ -1038,7 +1041,144 @@ createAddress:(body,addressId,userId)=>{
             console.log(result)
     })
     })
-}
+},
+findedProduct:(proId)=>{
+    return new Promise((resolve,reject)=>{
+        productInfo.findOne({_id:proId}).then((result)=>{
+            resolve(result)
+        })
+    })
+},
+getBuyNowProduct : (proId)=>{
+  return new Promise(async(resolve,reject)=>{
+     await productInfo.find({_id:proId},).lean().then((response)=>{
+          resolve(response)
+      })
+  })
+},
+addToWishlist:(proId, userId)=>{
+    return new Promise(async(resolve,reject)=>{
+        let proObj={
+            items : objectId(proId)
+        }
+        let userWishlist =await wishlistInfo.findOne({user:userId})
+       if(userWishlist){
+        let proExist = userWishlist.products.findIndex(e=> e.items==proId)
+        if(proExist=-1){
+            wishlistInfo.updateOne({user:userId},
+                {
+                    $push:{
+                        products:proObj
+                    }
+                }).then((response)=>{
+                    console.log("response;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;")
+                  console.log(response) 
+                    resolve(response)
+                })
+        }else{
+            reject(false)
+        }
+        } else{
+            const wishlist = new wishlistInfo({
+                user:userId,
+                products:proObj
+            })
+            wishlist.save((err,details)=>{
+                if(err){
+                    console.log("error"+err)
+                }else{
+                   
+                    resolve(details)
+                }
+            })
+        } 
+})
        
+
+},
+getWishlistProducts:(userId)=>{
+    return new Promise(async(resolve,reject)=>{
+        wishlistInfo.aggregate([
+            {
+                $match:{userId:userId}
+            },
+            {
+                $unwind:'$products'
+            },
+           
+            {
+                $lookup:{
+                    from:'products',
+                    localField:'products.items',
+                    foreignField:'_id',
+                    as:'wishlistproducts'
+                }
+            },
+            
+            {$unwind:"$wishlistproducts"}
+            ]).then((result)=>{
+                console.log("aggregated result +++++++++++++++++++++++++++++++")
+                console.log(result)
+                resolve(result)
+            })   
+    })
+},
+getWishlistCount:(userId)=>{
+    return new Promise(async(resolve,reject)=>{
+        let count = 0
+        let cart =await wishlistInfo.findOne({user:userId})
+        if(cart){
+            count=cart.products.length
+        }
+        resolve(count)
+    })
+},
+deleteItem:(id)=>{
+    return new Promise((resolve,rejection)=>{
+        let userId = id.userId
+        let proId = id.proId
+        wishlistInfo.updateOne({user:userId},{$pull:{products:{items:objectId(proId)}}}).then((res)=>{
+            resolve(true)
+        })
+    })
+ 
+},
+getweeklyreport: async () => {
+    const dayOfYear = (date) =>
+        Math.floor(
+            (date - new Date(date.getFullYear(), 0, 0)) / 1000 / 60 / 60 / 24
+        )
+    return new Promise(async (resolve, reject) => {
+       orderInfo.aggregate([
+            {
+                $match: {
+                   status: { $ne: 'cancelled' } ,
+                    date: { $gte: new Date(new Date() - 7 * 60 * 60 * 24 * 1000) },
+                },
+            },
+
+            { $group: { _id: { $dayOfYear: '$date' }, count: { $sum: 1 } } },
+        ]).then((result)=>{
+        const thisday = dayOfYear(new Date())
+        let salesOfLastWeekData = []
+        for (let i = 0; i < 8; i++) {
+            let count = result.find((d) => d._id === thisday + i - 7)
+
+            if (count) {
+                salesOfLastWeekData.push([`${i+1}`, count.count ])
+            } else {
+                salesOfLastWeekData.push([`${i+1}`, 0 ])
+            }
+        }
+        console.log("result ----------------------------")
+       console.log(salesOfLastWeekData)
+       
+        resolve(salesOfLastWeekData)
+        
+        })
+        
+
+    })
+}
 
 }
